Added Parts From Schematic
Some parts are still missing and there are a few I was unsure of the specs or how to find. Anything I was unsure of I put in the notes section. There were also a few parts that are out of stock so we might need to slightly change them or find them from a different supplier.